backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
scikit-learn==1.4.1.post1
xgboost==2.0.3
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4
